Base upload
This is the base of the program. It doesn't output how many times the account has been compromised and I plan on adding an email portion later to the program. I plan on adding more features in the future like showing what companies the email has been compromised on from the drop down.